refactor(bitwise.unit-test.ts): move rot tests from range-check.unit-test.ts to bitwise.unit-test.ts
